DCAC-153: Organise packages consistently by layer.
