+++ src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/ItemGroupStatusUpdaterApplicationTests.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.companieshouse.itemgroupstatusupdater.config.TestConfig;
 
 @SpringBootTest
 @Import(TestConfig.class)
