+++ src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/environment/EnvironmentVariablesCheckerTest.java
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import uk.gov.companieshouse.itemgroupstatusupdater.TestConfig;
+import uk.gov.companieshouse.itemgroupstatusupdater.config.TestConfig;
 
 @SpringBootTest
 @ActiveProfiles("test_main_positive")
